Verification before sending mail added

+++ src/main/java/pl/lodz/p/it/ssbd2024/ssbd01/mok/service/AccountService.java
@@ -225,6 +225,9 @@ public class AccountService {
         if (account.isEmpty()) {
             return;
         }
+        if (!account.get().getActive() || !account.get().getVerified()|| !account.get().isAccountNonLocked()) {
+            return;
+        }
         resetCredentialRepository.deleteByAccount(account.get());
         resetCredentialRepository.flush();
         CredentialReset credentialReset = verifier.saveTokenToResetCredential(account.get());
